assistant/requirements.txt
eel==0.14.0
PyInstaller==4.5.1
cchardet==2.1.7
qrcode==7.3.1
pillow==8.4.0